Add ByteBufUintUtil
This adds proper helpers to interface Uint types and ByteBuf. This
supersedes ByteBufWriteUtil.

Also migrate deprecated constructor/setter invocations in BGP/BMP,
which is made easier by the methods provided.

+++ bgp/extensions/linkstate/src/main/java/org/opendaylight/protocol/bgp/linkstate/impl/tlvs/MultiTopoIdTlvParser.java
@@ -35,7 +35,7 @@ public final class MultiTopoIdTlvParser implements LinkstateTlvParser<TopologyId
 
     @Override
     public TopologyIdentifier parseTlvBody(final ByteBuf value) {
-        return new TopologyIdentifier(value.readShort() & TlvUtil.TOPOLOGY_ID_OFFSET);
+        return new TopologyIdentifier(Uint16.valueOf(value.readShort() & TlvUtil.TOPOLOGY_ID_OFFSET));
     }
 
     @Override
